muita coisa kkk

+++ src/app/components/listagem-visitantes/listagem-visitantes.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 import { Visitante } from './visitante.model';
 import { Router } from '@angular/router';
 import { VisitanteService } from '@app/services/visitante-service.service';
@@ -6,43 +6,47 @@ import { VisitanteService } from '@app/services/visitante-service.service';
 
 @Component({
   selector: 'app-listagem-visitantes',
-  templateUrl: './listagem-visitantes.component.html',
-  styleUrls: ['./listagem-visitantes.component.scss']
+  templateUrl: './listagem-visitante.component.html',
+  styleUrls: ['./listagem-visitante.component.scss']
 })
-export class ListagemVisitasComponent implements OnInit {
+export class ListagemVisitantesComponent implements OnInit {
   visitantes: Visitante[] = [];
+  @Input() visitanteid = 0;
 
   constructor(
     private visitanteService: VisitanteService,
     private router: Router
-    ) {}
+  ) { }
 
   ngOnInit(): void {
     this.getVisitantes();
   }
 
-  getVisitantes(visitanteId:number = 0): void {
-    this.visitanteService.getVisitante()
+  getVisitantes(): void {
+    this.visitanteService.
+    getAllVisitantesByVisitaID(this.visitanteid)
       .subscribe((visitantes: Visitante[]) => {
         console.log(visitantes)
         this.visitantes = visitantes;
       });
   }
 
-  excluirVisita(visitanteId:number): void {
-    this.visitanteService.excluirVisitante(visitanteId).subscribe(
-      () => {
-        console.log('Visitante excluida com sucesso');
-        this.getVisitantes(visitanteId);
-      },
-      (error) => {
-        console.error('erro ao deletar a visitante', error);
-      }
-    );
+  excluirVisitante(visitanteId: number): void {
+    if (confirm('Tem certeza que deseja EXCLUIR este visitante?')) {
+      this.visitanteService.excluirVisitante(visitanteId).subscribe(
+        () => {
+          console.log('Visitante excluida com sucesso');
+          this.getVisitantes();
+        },
+        (error) => {
+          console.error('erro ao deletar a visitante', error);
+        }
+      );
+    }
   }
-  
+
   editarVisitante(visitanteId: number): void {
-    this.router.navigateByUrl(`/cadastro?id=${visitanteId}`);
+    this.router.navigateByUrl(`/cadastroVisitante?id=${visitanteId}`);
   }
-  
+
 }
